+++ openmp/exactopenmp.h
